River-container name has been removed from the project folders.
Background:

As a result of this thread (http://article.gmane.org/gmane.comp.apache.incubator.river.devel/9626),
development of the River Container has moved to Github, rather than continue at
Apache.  As such, the names of all the components are now reverted to the
Harvester Application Container, as begun at StratusCom in the early 2000's.